+++ fs/types/bigint/proof.f.ts
@@ -10,7 +10,9 @@ import {
     divUp,
     roundUp,
     divUpE2,
-    roundUpE2
+    roundUpE2,
+    divUp8,
+    roundUp8
 } from './module.f.ts'
 import { assertEq } from '../../asserts/module.f.ts'
 import { min } from '../function/compare/module.f.ts'
@@ -413,5 +415,17 @@ export const proof = {
         assertEq(roundUpE2(3n)(0n), 0n)
         assertEq(roundUpE2(3n)(8n), 8n)
         assertEq(roundUpE2(3n)(9n), 16n)
+    },
+    divUp8: () => {
+        assertEq(divUp8(0n), 0n)
+        assertEq(divUp8(1n), 1n)
+        assertEq(divUp8(8n), 1n)
+        assertEq(divUp8(9n), 2n)
+    },
+    roundUp8: () => {
+        assertEq(roundUp8(0n), 0n)
+        assertEq(roundUp8(1n), 8n)
+        assertEq(roundUp8(8n), 8n)
+        assertEq(roundUp8(9n), 16n)
     },
 }